--- src/main/java/balancetalk/domain/Notice.java
@@ -1,37 +1,0 @@
-package balancetalk.domain;
-
-import balancetalk.domain.file.entity.File;
-import balancetalk.domain.member.entity.Member;
-import balancetalk.global.common.BaseTimeEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import java.util.ArrayList;
-import java.util.List;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Notice extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "notice_id")
-    private Long id;
-
-    private String title;
-    private String content;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @OneToMany(mappedBy = "notice")
-    private List<File> files = new ArrayList<>();
-}
