--- src/main/java/balancetalk/domain/comment/entity/Comment.java
@@ -1,50 +1,0 @@
-package balancetalk.domain.comment.entity;
-
-import balancetalk.domain.report.entity.Report;
-import balancetalk.domain.ViewStatus;
-import balancetalk.domain.member.entity.Member;
-import balancetalk.domain.post.entity.Post;
-import balancetalk.global.common.BaseTimeEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import java.util.ArrayList;
-import java.util.List;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Comment extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "comment_id")
-    private Long id;
-
-    private String content;
-
-    @Enumerated(value = EnumType.STRING)
-    private ViewStatus viewStatus;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
-
-    @OneToMany(mappedBy = "comment")
-    private List<CommentLike> likes = new ArrayList<>();
-
-    @OneToMany(mappedBy = "comment")
-    private List<Report> reports = new ArrayList<>();
-}
