--- src/main/java/balancetalk/domain/file/enums/FileType.java
@@ -1,5 +1,0 @@
-package balancetalk.domain.file.enums;
-
-public enum FileType {
-    JPG, JPEG, PNG
-}
