--- src/main/java/balancetalk/domain/member/enums/Role.java
@@ -1,5 +1,0 @@
-package balancetalk.domain.member.enums;
-
-public enum Role {
-    ADMIN, USER, GUEST
-}
