--- src/main/java/balancetalk/domain/report/enums/ReportCategory.java
@@ -1,6 +1,0 @@
-package balancetalk.domain.report.enums;
-
-public enum ReportCategory {
-
-    COMMERCIAL, ILLEGAL, ABUSE, ETC
-}
